+++ imagemeta.go
@@ -13,6 +13,7 @@ import (
 	"github.com/evanoberholster/imagemeta/isobmff"
 	"github.com/evanoberholster/imagemeta/jpeg"
 	"github.com/evanoberholster/imagemeta/meta"
+	"github.com/evanoberholster/imagemeta/png"
 	"github.com/evanoberholster/imagemeta/tiff"
 	"github.com/pkg/errors"
 )
@@ -85,8 +86,8 @@ func Decode(r io.ReadSeeker) (exif2.Exif, error) {
 // DecodeCR3 decodes a CR3 file from an io.Reader returning Exif or an error.
 func DecodeCR3(r io.ReadSeeker) (exif2.Exif, error) {
 	rr := readerPool.Get().(*bufio.Reader)
-	rr.Reset(r)
 	defer readerPool.Put(rr)
+	rr.Reset(r)
 
 	ir := exif2.NewIfdReader(exif2.Logger)
 	defer ir.Close()
@@ -103,10 +104,7 @@ func DecodeCR3(r io.ReadSeeker) (exif2.Exif, error) {
 	if err := bmr.ReadMetadata(); err != nil {
 		return ir.Exif, err
 	}
-	// Set ImageType to CR3
-	//ir.Exif.ImageType = imagetype.ImageCR3
 	return ir.Exif, nil
-	//return exif2.DecodeHeader(r, moov.Meta.Exif[0], moov.Meta.Exif[1], moov.Meta.Exif[3])
 }
 
 // DecodeTiff decodes a Tiff/DNG file from an io.Reader returning Exif or an error.
@@ -136,4 +134,51 @@ func DecodeTiff(r io.ReadSeeker) (exif2.Exif, error) {
 // DecodeCR2 decodes a CR2 file from an io.Reader returning Exif or an error.
 func DecodeCR2(r io.ReadSeeker) (exif2.Exif, error) {
 	return DecodeTiff(r)
+}
+
+// DecodeHeif decodes a Heif file from an io.Reader returning Exif or an error.
+// Needs improvement
+func DecodeHeif(r io.ReadSeeker) (exif2.Exif, error) {
+	return DecodeTiff(r)
+}
+
+// DecodeJPEG decodes a JPEG file from an io.Reader returning Exif or an error.
+func DecodeJPEG(r io.ReadSeeker) (exif2.Exif, error) {
+	rr := readerPool.Get().(*bufio.Reader)
+	rr.Reset(r)
+	defer readerPool.Put(rr)
+
+	ir := exif2.NewIfdReader(exif2.Logger)
+	defer ir.Close()
+
+	it, err := imagetype.ScanBuf(rr)
+	if err != nil {
+		return exif2.Exif{}, err
+	}
+	if it != imagetype.ImageJPEG {
+		return exif2.Exif{}, ErrMetadataNotSupported
+	}
+
+	if err = jpeg.ScanJPEG(rr, ir.DecodeJPEGIfd, nil); err != nil {
+		return exif2.Exif{}, err
+	}
+	ir.Exif.ImageType = it
+	return ir.Exif, nil
+}
+
+// DecodePng decodes a PNG file from an io.Reader returning Exif or an error.
+func DecodePng(r io.ReadSeeker) (exif2.Exif, error) {
+	header, err := png.ScanPngHeader(r)
+	if err != nil {
+		return exif2.Exif{}, err
+	}
+
+	ir := exif2.NewIfdReader(exif2.Logger)
+	defer ir.Close()
+
+	if err := ir.DecodeTiff(r, header); err != nil {
+		return ir.Exif, err
+	}
+
+	return ir.Exif, nil
 }
